backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
openai
faiss-cpu    # if on Linux/amd64, otherwise use faiss-cpu prebuilt wheel for your platform
numpy
tqdm
pypdf        # for PDF extraction
python-dotenv
requests